Database check 1

+++ src/main/java/func/PetitionRepository.java
@@ -1,10 +1,13 @@
 package func;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 
 @Repository
-public interface PetitionRepository extends JpaRepository<Petitie, Long> {
+public interface PetitionRepository extends JpaRepository<Petitie, Integer> {
 
 }
